LPD-19398 SimilarResultsContributorsRegistry is an internal interface, it should not be registered as osgi service
